+++ test/unit/legacy_unit/util/copy.c
@@ -3,9 +3,9 @@
 #define F "/etc/services"
 
 int main () {
-	int len;
+	size_t len;
 	char *out= r_file_slurp (F, &len);
-	r_file_dump ("a", out, len);
+	r_file_dump ("a", out, (int)len);
 	system ("md5 "F);
 	system ("md5 a");
 	return 0;
